server/requirements.txt
uvicorn[standard]
fastapi
bcrypt
python-jose[cryptography]
pydantic[email]
sqlalchemy
psycopg2-binary
python-multipart
passlib
black